*) Some minor stuff, most notably solved the free memory overflow by changing the StatusVariable to long type
git-svn-id: svn://svn.paxle.net/paxle/trunk@2614 f5d3ab42-bacf-4b1a-9692-565be65a0cfe

+++ bundles/Core/src/main/java/org/paxle/core/monitorable/provider/impl/RuntimeMemoryMonitoring.java
@@ -95,7 +95,7 @@ public class RuntimeMemoryMonitoring implements Monitorable {
 		return new StatusVariable(
 				name,
 				StatusVariable.CM_GAUGE,
-				(int)mem
+				mem
 		);
 	}
